+++ src/render/index.html
@@ -5,7 +5,7 @@
     <title>Ease Video Player</title>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link href="../node_modules/purecss/build/pure-min.css" rel="stylesheet">
+    <link href="../../node_modules/purecss/build/pure-min.css" rel="stylesheet">
     <link href="style/main.css" rel="stylesheet">
     <link href="style/index.css" rel="stylesheet">
 </head>
@@ -15,11 +15,11 @@
     <div id="app"></div>
     
     <!-- Dependencies -->
-    <script src="../node_modules/react/dist/react.js"></script>
-    <script src="../node_modules/react-dom/dist/react-dom.js"></script>
+    <script src="../../node_modules/react/dist/react.js"></script>
+    <script src="../../node_modules/react-dom/dist/react-dom.js"></script>
 
     <!-- Webpack file -->
-    <script src="../dist/bundle.js"></script>
+    <script src="../../dist/bundle.js"></script>
 </body>
 
 </html>
