+++ src/render/style/index.css
@@ -1,0 +1,1 @@
+@font-face{font-family:'Indie Flower';font-style:cursive;src:url('font/IndieFlower.ttf')}@font-face{font-family:'Inconsolata';font-style:normal;src:url('font/Inconsolata-Regular.ttf')}#center{width:50%;margin:0 auto}#indie-flower{font-family:'Indie Flower',cursive}#inconsolata{font-family:'Inconsolata',regular}body{font-family:'Indie Flower',cursive;background-color:#303030}body>input [type='number']{-moz-appearance:textfield}body>input[type="number"]::-webkit-outer-spin-button,body input[type="number"]::-webkit-inner-spin-button{-webkit-appearance:none;margin:0}body html,body button,body input,body select,body textarea,body .pure-g [class*="pure-u"]{font-family:'Indie Flower',cursive}body .join{text-align:center}body .join input{font-family:'Inconsolata',regular;text-align:center;background-color:#303030;font-size:18px;padding:10px 10px 10px 5px;width:200px;border:none;border-bottom:1px solid black}body .join input:focus{outline:none}body .break{text-align:center;margin-top:30px}body .break .or{font-size:1.5em}body .break .line{margin-top:16px;border-top:1px solid black}body .host{text-align:center}body .host .upload-box{height:300px;width:400px;position:relative;display:block;margin:0 auto;border-radius:10px;border:1px dashed black}body .host .upload-box>.upload-here{position:relative;display:inline-block;padding-top:125px;padding-bottom:125px}
