+++ src/render/style/main.css
@@ -1,0 +1,1 @@
+@font-face{font-family:'Indie Flower';font-style:cursive;src:url('font/IndieFlower.ttf')}@font-face{font-family:'Inconsolata';font-style:normal;src:url('font/Inconsolata-Regular.ttf')}#center{width:50%;margin:0 auto}#indie-flower{font-family:'Indie Flower',cursive}#inconsolata{font-family:'Inconsolata',regular}body{font-family:'Indie Flower',cursive;background-color:#303030}body>input [type='number']{-moz-appearance:textfield}body>input[type="number"]::-webkit-outer-spin-button,body input[type="number"]::-webkit-inner-spin-button{-webkit-appearance:none;margin:0}body html,body button,body input,body select,body textarea,body .pure-g [class*="pure-u"]{font-family:'Indie Flower',cursive}
