+++ src/render/video/VideoClientPage.tsx
