+++ test/main/MainTest.ts
@@ -2,7 +2,7 @@ import * as Assert from "assert";
 
 import { app, BrowserWindow } from "electron";
 
-import Main from "../../src/Main";
+import Main from "../../src/main/Main";
 
 describe("Main", () => {
     describe("#fakeTest", () => {
